features/car_parts_features/backlight_feature/requirements.txt
torch==1.8.0
torchvision==0.9.0
numpy
Pillow